Rename new points files with originals' names
Now the originals have been removed
